feat(implementacao-jwt-refactor): refatora codigo para separar corretamente os dominios de restaurant user e customer user e implementa jwt token para ambos os tipos de usuario

+++ src/main/java/com/eatz/presentation/web/customer/dto/CustomerResponse.java
@@ -2,15 +2,13 @@ package com.eatz.presentation.web.customer.dto;
 
 import com.eatz.presentation.web.address.dto.AddressResponse;
 
-import java.util.UUID;
-
 public class CustomerResponse {
     private String name;
     private String email;
     private String password;
     private AddressResponse address;
 
-    public CustomerResponse(Integer id, String name, String email, AddressResponse endereco) {
+    public CustomerResponse(Long id, String name, String email, AddressResponse endereco) {
         this.name = name;
         this.email = email;
         this.address = endereco;
